billmap/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.3
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.29
alembic>=1.13.1

# PDF processing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.3.0

# LLM providers
anthropic>=0.25.0
openai>=1.14.0

# QuickBooks Online
python-quickbooks>=0.9.7

# Utilities
python-dotenv>=1.0.1
python-Levenshtein>=0.25.0
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
